+++ KazLibrary/Loader/ObjResourceMgr.h
@@ -48,7 +48,7 @@ struct ObjResourceData
 
 	D3D12_VERTEX_BUFFER_VIEW vertexBufferView;
 	D3D12_INDEX_BUFFER_VIEW indexBufferView;
-	unsigned short indexNum;
+	UINT indexNum;
 };
